ejemplo14

+++ src/es/avalon/Ejemplo014Media.java
@@ -4,7 +4,7 @@ import javax.sound.midi.Soundbank;
 
 public class Ejemplo014Media {
     public static void main(String[] args) {
-        double[] lista = new double[]{7, 10, 2, 5, 3, 0};
+        int[] lista = new int[]{7, 10, 2, 5, 3, 0};
         double suma = 0;
         for (int i = 0; i < lista.length; i++) {
             suma = suma + lista[i];
